crud event

+++ spring/src/main/java/com/example/spring/persistance/repositories/EvenementRepository.java
@@ -3,5 +3,5 @@ package com.example.spring.persistance.repositories;
 import com.example.spring.persistance.entities.Evenement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EvenementRepository extends JpaRepository<Evenement,Integer> {
+public interface EvenementRepository extends JpaRepository<Evenement,Long> {
 }
