+++ spring/src/main/java/com/example/spring/persistance/repositories/FeedbackRepository.java
@@ -3,5 +3,5 @@ package com.example.spring.persistance.repositories;
 import com.example.spring.persistance.entities.Feedback;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeedbackRepository extends JpaRepository<Feedback,Integer>{
+public interface FeedbackRepository extends JpaRepository<Feedback,Long>{
 }
