+++ spring/src/main/java/com/example/spring/persistance/repositories/FormationRepository.java
@@ -3,5 +3,5 @@ package com.example.spring.persistance.repositories;
 import com.example.spring.persistance.entities.Formation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FormationRepository extends JpaRepository<Formation,Integer> {
+public interface FormationRepository extends JpaRepository<Formation,Long> {
 }
