+++ spring/src/main/java/com/example/spring/persistance/repositories/ReclamationRepository.java
@@ -3,5 +3,5 @@ package com.example.spring.persistance.repositories;
 import com.example.spring.persistance.entities.Reclamation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReclamationRepository extends JpaRepository<Reclamation,Integer> {
+public interface ReclamationRepository extends JpaRepository<Reclamation,Long> {
 }
